+++ lib/core/flame/components/entities/mobs/rat.dart
@@ -12,17 +12,17 @@ class Rat extends DraggableEntity with DisappearOnDeath {
     entityResourceName: 'rat',
     defaultSize: Vector2(32, 32),
     defaultScale: 2,
-    walkingConfig: AnimationConfig(frames: 6, stepTime: 0.12),
+    walkingConfig: AnimationConfig(frames: 6, stepTime: 0.24),
     attackingConfig: AnimationConfig(frames: 6, stepTime: 0.2),
     dragConfig: AnimationConfig(frames: 6, stepTime: 0.2),
-    dyingConfig: AnimationConfig(frames: 6, stepTime: 0.12),
+    dyingConfig: AnimationConfig(frames: 6, stepTime: 0.2),
     damageOnAttack: 5,
     goldOnKill: 5,
-    baseWalkingSpeed: 42,
+    baseWalkingSpeed: 55,
   );
 
   late final RectangleHitbox _hitBox =
-      EntityHelper.createRectangleHitbox(size: Vector2(20, 15), position: Vector2(8, 16), anchor: Anchor.bottomCenter);
+      EntityHelper.createRectangleHitbox(size: Vector2(18, 12), position: Vector2(16, 32), anchor: Anchor.bottomCenter);
 
   Rat({super.scaleModifier, super.modifiedWalkingSpeed}) : super(entityConfig: _ratConfig) {
     setDisappearSpeedFactor(1.5);
@@ -30,7 +30,7 @@ class Rat extends DraggableEntity with DisappearOnDeath {
 
   @override
   Vector2? attackEffectPosition() {
-    return position + Vector2(scaledSize.x - 15, -scaledSize.y / 2);
+    return position + Vector2(scaledSize.x, -scaledSize.y / 2);
   }
 
   @override
@@ -43,7 +43,7 @@ class Rat extends DraggableEntity with DisappearOnDeath {
     final rat = Rat(scaleModifier: scaleModifier, modifiedWalkingSpeed: _ratConfig.baseWalkingSpeed * speedFactor);
 
     // Since we are using the bottom left anchor, we need to adjust the position.
-    rat.position = position + Vector2(-rat.scaledSize.x, rat.scaledSize.y);
+    rat.position = position - Vector2(rat.scaledSize.x, rat.scaledSize.y / 2);
     return rat;
   }
 }